% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_burn_subtitles.R
\name{ari_burn_subtitles}
\alias{ari_burn_subtitles}
\title{Burn Subtitles into a video}
\usage{
ari_burn_subtitles(video, srt, verbose = FALSE)
}
\arguments{
\item{video}{Video in \code{mp4} format}

\item{srt}{Subtitle file in \code{srt} format}

\item{verbose}{print diagnostic messages.  If > 1,
then more are printed}
}
\value{
Name of output video
}
\description{
Burn Subtitles into a video
}
\note{
This needs \code{ffmpeg} that was compiled with
\code{--enable-libass} as per
\url{https://trac.ffmpeg.org/wiki/HowToBurnSubtitlesIntoVideo}
}
