% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_example.R
\name{ari_example}
\alias{ari_example}
\title{Get the path to an ari example file}
\usage{
ari_example(path = NULL)
}
\arguments{
\item{path}{The name of the file. If no argument is provided then
all of the example files will be listed.}
}
\description{
This function allows you to quickly access files that are used in
the ari documentation.
}
\examples{
ari_example("ari_intro.Rmd")
}

