% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_narrate.R
\name{ari_narrate}
\alias{ari_narrate}
\title{Create a video from slides and a script}
\usage{
ari_narrate(script, slides, output = "output.mp4", voice, ws_args = list())
}
\arguments{
\item{script}{A markdown file where every paragraph will be read over a
corresponding slide.}

\item{slides}{A URL for an HTML slideshow created with \code{rmarkdown},
\code{xaringan}, or a similar package.}

\item{output}{A path to the video file which will be created.}

\item{voice}{The Amazon Polly voice you want to use. See \code{list_voices}.}

\item{ws_args}{Arguments to be passed to \code{webshot}.}
}
\description{
Given a script and slides, make a video.
}

