% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_subtitles.R
\name{ari_subtitles}
\alias{ari_subtitles}
\title{Generate subtitle files for audio for video content}
\usage{
ari_subtitles(paragraphs, wavs, path, width = 42)
}
\arguments{
\item{paragraphs}{String of text}

\item{wavs}{Wave objects from tuneR}

\item{path}{Path to .srt file output}

\item{width}{Width of each subtitle}
}
\description{
Generate subtitle files for audio for video content
}
\examples{
\dontrun{
# Audio files (.wav)
audio <- system.file("test", c("audio1.wav", "audio2.wav"), package = "ari")

# Import audio files as wave objects
first_wav_object <- tuneR::readWave(audio[1])
second_wav_object <- tuneR::readWave(audio[2])
# Put wave objects inside a list
wavs <- list(first_wav_object, second_wav_object)

# Text
sentences <- c(
  "Welcome to my very interesting lecture.",
  "Here are some fantastic equations I came up with."
)

ari_subtitles(sentences, wavs, "output.srt")
}
}
