% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_talk.R
\name{ari_talk}
\alias{ari_talk}
\title{Create spoken audio files}
\usage{
ari_talk(paragraphs, output = "output.wav", voice = "Joanna",
  service = "amazon")
}
\arguments{
\item{paragraphs}{A vector strings that will be spoken by Amazon Polly.}

\item{output}{A path to the audio file which will be created.}

\item{voice}{The  voice you want to use. See 
\code{\link[text2speech]{tts_voices}} for more information 
about what voices are available.}

\item{service}{speech synthesis service to use,
passed to \code{\link[text2speech]{tts}}}
}
\description{
A simple function for demoing how spoken text will sound.
}
