% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coqui_args.R
\name{coqui_args}
\alias{coqui_args}
\title{List of arguments to the Coqui text-to-speech engine}
\usage{
coqui_args(model_name = "jenny", vocoder_name = "jenny")
}
\arguments{
\item{model_name}{Deep Learning model for Text-to-Speech Conversion}

\item{vocoder_name}{Voice coder used for speech coding and transmission}
}
\value{
List of arguments
}
\description{
List of arguments to the Coqui text-to-speech engine
}
\examples{
coqui_args(model_name = "tacotron2-DDC_ph", vocoder_name = "ljspeech/univnet")
}
