% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gs_file.R
\name{download_gs_file}
\alias{download_gs_file}
\title{Download Google Slides File}
\usage{
download_gs_file(gs_url, out_type = "pptx")
}
\arguments{
\item{gs_url}{Link to Google slides presentation, passed to
\code{\link{get_slide_id}}}

\item{out_type}{output type of file to download. Usually
`pdf` or `pptx`}
}
\value{
Downloaded file (in temporary directory)
}
\description{
Download Google Slides File
}
\note{
This downloads presentations if they are public and also try to make
sure it does not fail on large files
}
