% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_codecs.R
\name{ffmpeg_codecs}
\alias{ffmpeg_codecs}
\alias{ffmpeg_video_codecs}
\alias{ffmpeg_audio_codecs}
\alias{ffmpeg_muxers}
\title{Get Codecs for ffmpeg}
\usage{
ffmpeg_codecs()

ffmpeg_video_codecs()

ffmpeg_audio_codecs()

ffmpeg_muxers()
}
\value{
A `data.frame` of codec names and capabilities
}
\description{
Get Codecs for ffmpeg
}
\examples{
ffmpeg_codecs()
ffmpeg_audio_codecs()
ffmpeg_video_codecs()
}
