% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_convert.R
\name{ffmpeg_convert}
\alias{ffmpeg_convert}
\title{Convert Files using FFMPEG}
\usage{
ffmpeg_convert(
  file,
  outfile = tempfile(fileext = paste0(".", tools::file_ext(file))),
  overwrite = TRUE,
  args = NULL
)
}
\arguments{
\item{file}{Video/PNG file to convert}

\item{outfile}{output file}

\item{overwrite}{should output file be overwritten?}

\item{args}{arguments to pass to \code{\link{system2}} to pass to
\code{ffmpeg}}
}
\value{
A character string of the output file with different attributes
}
\description{
Convert Files using FFMPEG
}
\examples{
pngfile <- tempfile(fileext = ".png")
png(pngfile)
plot(0, 0)
dev.off()
if (have_ffmpeg_exec()) {
  res <- ffmpeg_convert(pngfile)
}
}
