% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_exec.R
\name{ffmpeg_error_log}
\alias{ffmpeg_error_log}
\title{Check error output from individual video}
\usage{
ffmpeg_error_log(file, verbose = TRUE)
}
\arguments{
\item{file}{path to video}

\item{verbose}{print diagnostic messages}
}
\value{
The output of the error log
}
\description{
Check error output from individual video
}
