% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_exec.R
\name{ffmpeg_exec}
\alias{ffmpeg_exec}
\alias{have_ffmpeg_exec}
\title{Get Path to ffmpeg Executable}
\usage{
ffmpeg_exec()

have_ffmpeg_exec()
}
\value{
The path to the \code{ffmpeg} executable, or an error.
}
\description{
Get Path to ffmpeg Executable
}
\examples{
\dontrun{
if (have_ffmpeg_exec()) { 
ffmpeg_exec()
}
}
}
