% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.curve.plot}
\alias{get.AT.curve.plot}
\title{This function creates a plot of the signals generated with an
Aaslid-Tiecks (A-T) model.
Specifically:
- plots curve$ABP.normalised and curve$CBFV.step.response signals using
  curve$time.instants as x-axis
- annotates the plot with the values of curve$T, curve$D, curve$K and
  curve$sampling.time
- annotates the plot with a vertical grey line marking the time of
  thigh-cuff release
- uses the argument title (if specified) as the plot's title
The function receives the following arguments:}
\usage{
get.AT.curve.plot(curve, title = NULL)
}
\arguments{
\item{curve:}{a list with the A-T curve, as returned by the function
get.AT.curve()}

\item{title:}{the title for the plot. Default value: NULL}
}
\value{
The function returns the resulting plot
}
\description{
This function creates a plot of the signals generated with an
Aaslid-Tiecks (A-T) model.
Specifically:
- plots curve$ABP.normalised and curve$CBFV.step.response signals using
  curve$time.instants as x-axis
- annotates the plot with the values of curve$T, curve$D, curve$K and
  curve$sampling.time
- annotates the plot with a vertical grey line marking the time of
  thigh-cuff release
- uses the argument title (if specified) as the plot's title
The function receives the following arguments:
}

