% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.curve}
\alias{get.AT.curve}
\title{This function creates a normalised ABP step stimulus and obtains the
CBFV response according to the Aaslid-Tiecks (A-T) model with a
specific level of efficiency.}
\usage{
get.AT.curve(sampling.time = 0.1, time.until.release = 10,
  time.after.release = 20, smooth.step.stimulus = FALSE, filter.order = 2,
  cutoff.frequency = 0.2,
  left.stabilisation.time = ifelse(smooth.step.stimulus, 30, 0),
  time.rounding.digits = format.info(sampling.time)[2], T = 1.9, D = 0.75,
  K = 0.9, stabilisation.time = 1)
}
\arguments{
\item{time.rounding.digits:}{the parameters required to generate a normalised
ABP stimulus.
Default values: as in function
get.normalised.ABP.stimulus()}

\item{T:}{Parámetro del modelo correspondiente al tiempo. El valor por defecto es 1.9}

\item{D:}{Parámetro del modelo correspondiente al factor de amortiguamiento. El valor por defecto es 0.75}

\item{K:}{Parámetro del modelo correspondiente a la ganancia dinámica autorreguladora. El valor por defecto es 0.9}

\item{stabilisation.time:}{tiempo de estabilización de la autorregulación. Por defecto el valor es 1.}
}
\value{
The function's answer is a list with the following:
.$T, .$D, .$K: the parameters of the model used
.$time.instants: the specified time instants.
.$sampling.time: the sampling time of the signals.
.$ABP.normalised: the specified normalised ABP stimulus.
.$CBFV.step.response: the CBFV response given by the model to the
                      specified ABP stimulus.
.$time.release: the time instant in which the thigh-cuffs are
                supposedly released
}
\description{
T, D, K, stabilisation.time: the parameters necessary to obtain the
                             CBFV response to the ABP stimulus
                             generated.
                             Default values: as in function
                             get.theoretical.CBFV.response()
}

