% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.curves.plot}
\alias{get.AT.curves.plot}
\title{This function creates a plot with CBFV responses generated by
Aaslid-Tiecks (A-T) models  (normally for the same ABP stimulus).
The function receives the following argument:}
\usage{
get.AT.curves.plot(curves, label.name = NULL, labels = NULL, title = NULL)
}
\arguments{
\item{curves:}{a list in which each element is a list with an A-T curve as
returned by the function get.AT.curve()}

\item{label.name:}{the name for the legend of the plot. If NULL, no legend is
added and curves are plotted in black.
Default value: NULL}

\item{labels:}{a vector of the same length as curves, containing the label for
each curve. If label.name is not NULL, these labels are used as
the colour aesthetic and for labelling the curves in the legend.
Default value: NULL}

\item{title:}{the title for the plot.
Default value: NULL}
}
\value{
The function returns the resulting plot
}
\description{
This function creates a plot with CBFV responses generated by
Aaslid-Tiecks (A-T) models  (normally for the same ABP stimulus).
The function receives the following argument:
}

