% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.decimal.templates.parameters}
\alias{get.AT.decimal.templates.parameters}
\title{Se genera los parámetros para el modelo AT correspondientes a los valores decimales del ARI.
Desde el 0 hasta el 9 en intervalos de 0.1.}
\usage{
get.AT.decimal.templates.parameters(rounding.digits = 6)
}
\arguments{
\item{rounding.digits:}{Cantidad de valores decimales a considerar.}
}
\description{
Se genera los parámetros para el modelo AT correspondientes a los valores decimales del ARI.
Desde el 0 hasta el 9 en intervalos de 0.1.
}

