% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.decimal.templates}
\alias{get.AT.decimal.templates}
\title{se obtiene 91 plantillas}
\usage{
get.AT.decimal.templates(sampling.time = 0.1, time.until.release = 10,
  time.after.release = 20, smooth.step.stimulus = FALSE, filter.order = 2,
  cutoff.frequency = 0.2,
  left.stabilisation.time = ifelse(smooth.step.stimulus, 30, 0),
  time.rounding.digits = format.info(sampling.time)[2],
  stabilisation.time = 1, rounding.digits = 6)
}
\description{
se obtiene 91 plantillas
}

