% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.templates.parameters}
\alias{get.AT.templates.parameters}
\title{Obtención de los parámetros T, D y K para los ARI desde el 0 al 9.}
\usage{
get.AT.templates.parameters()
}
\value{
Data frame con los parámetros asociados a cada ARI.
}
\description{
Obtención de los parámetros T, D y K para los ARI desde el 0 al 9.
}

