% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.AT.templates}
\alias{get.AT.templates}
\title{Se generan los templates utilizados para el cálculo del ARI según el modelo de AT. Estos templates son 9
números enteros.}
\usage{
get.AT.templates(sampling.time = 0.1, time.until.release = 10,
  time.after.release = 20, smooth.step.stimulus = FALSE, filter.order = 2,
  cutoff.frequency = 0.2,
  left.stabilisation.time = ifelse(smooth.step.stimulus, 30, 0),
  time.rounding.digits = format.info(sampling.time)[2],
  stabilisation.time = 1)
}
\value{
Data frame con los siguientes datos:
       Los parámetros del modelo AT: $T, $D y $K.
       $time.instants: instantes de tiempo del monitoreo.
       $sampling.time: tiempo de muestreo.
       $ABP.normalised: Estímulo en la presión normalizado.
       $CBFV.step.response: Respuesta de la velocidad de flujo cerebral al cambio de la presión.
       $time.release: Instante de tiempo en que se aplica la maniobra.
       $ARI.value: Valor del índice ARI.
}
\description{
Se generan los templates utilizados para el cálculo del ARI según el modelo de AT. Estos templates son 9
números enteros.
}

