% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atARI-v15.8.15.R
\name{get.theoretical.CBFV.response}
\alias{get.theoretical.CBFV.response}
\title{This function obtains the CBFV response to an ABP stimulus
according to the Aaslid-Tiecks (A-T) model with the specified
parameters.
T, D and K: the model parameters.
            Default value: values that generates a response with
            A-T ARI = 5.}
\usage{
get.theoretical.CBFV.response(T = 1.9, D = 0.75, K = 0.9, time.instants,
  ABP.normalised, sampling.time = min(round(diff(time.instants), 3)),
  stabilisation.time = 1)
}
\arguments{
\item{T:}{Par치metro del modelo correspondiente al tiempo. El valor por defecto es 1.9}

\item{D:}{Par치metro del modelo correspondiente al factor de amortiguamiento. El valor por defecto es 0.75}

\item{K:}{Par치metro del modelo correspondiente a la ganancia din치mica autorreguladora. El valor por defecto es 0.9}

\item{time.instants:}{the time instants in which the input ABP stimulus was
sampled.}

\item{ABP.normalised:}{the normalised ABP stimulus to be used as input to the
A-T model.}

\item{sampling.time:}{the sampling time used in the input ABP stimulus.
Default value: the smallest difference between time
instants (rounded to 3 decimal places - max. 1000 Hz).}

\item{stabilisation.time:}{the time to be used for stabilisation of the model,
before and after the input stimulus.
Default value: 1 s}
}
\value{
The function's answer is a list with the following:
.$T, .$D, .$K: the parameters of the model used
.$time.instants: a copy of the specified time instants.
.$sampling.time: a copy of the specified sampling time of the signals.
.$ABP.normalised: the specified normalised ABP stimulus.
.$CBFV.theoretical.response: the CBFV response given by the A-T model
                             with the specified parameters to the given
                             ABP stimulus.
}
\description{
This function obtains the CBFV response to an ABP stimulus
according to the Aaslid-Tiecks (A-T) model with the specified
parameters.
T, D and K: the model parameters.
            Default value: values that generates a response with
            A-T ARI = 5.
}
\references{
Frank P. Tiecks, Arthur M. Lam, Rune Aaslid, David W. Newell. Comparison of Static and Dynamic Cerebral Autoregulation Measurements.
Stroke, 1995, 26, pages 1014-1019.
}

