% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_wav.R
\name{pad_wav}
\alias{pad_wav}
\title{Pad Wave Objects}
\usage{
pad_wav(wav, duration = NULL)
}
\arguments{
\item{wav}{list of Wave objects}

\item{duration}{If \code{NULL}, the duration will simply round
the Wave up to the next whole integer.  If not, these are the
duration to pad the Wave *to*. For example 12 means the output
Wave will have a length of 12 seconds.  Pass \code{NA} to those
Waves that you want simple rounding.}
}
\value{
A list of Wave objects, same length as input \code{wav}
}
\description{
Pad Wave Objects
}
\examples{
wavs <- list(
  tuneR::noise(duration = 1.85 * 44100),
  tuneR::noise()
)
out <- pad_wav(wavs)
dur <- sapply(out, function(x) length(x@left) / x@samp.rate)
duration <- c(2, 2)
out <- pad_wav(wavs, duration = duration)
dur <- sapply(out, function(x) length(x@left) / x@samp.rate)
stopifnot(all(dur == duration))
duration <- c(2, 2.5)
out <- pad_wav(wavs, duration = duration)
dur <- sapply(out, function(x) length(x@left) / x@samp.rate)
stopifnot(isTRUE(all.equal(dur, duration)))
}
