% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_notes.R
\name{pptx_notes}
\alias{pptx_notes}
\alias{pptx_slide_text_df}
\alias{pptx_slide_note_df}
\alias{unzip_pptx}
\title{Get Notes from a PowerPoint (usually from Google Slides)}
\usage{
pptx_notes(file, ...)

pptx_slide_text_df(file, ...)

pptx_slide_note_df(file, ...)

unzip_pptx(file)
}
\arguments{
\item{file}{Character. Path for `PPTX` file}

\item{...}{additional arguments to pass to \code{\link{xml_notes}},
particularly \code{xpath}}
}
\value{
Either a character vector or `NULL`
}
\description{
Get Notes from a PowerPoint (usually from Google Slides)
}
\examples{
ex_file = system.file("extdata", "example.pptx",
package = "ariExtra")
pptx_notes(ex_file)
pptx_slide_note_df(ex_file)
pptx_slide_text_df(ex_file)
}
