% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gs_file.R
\name{pptx_to_pdf}
\alias{pptx_to_pdf}
\title{Convert a PPTX file to a PDF file}
\usage{
pptx_to_pdf(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to the PPTX file that needs to be converted to PDF.}

\item{verbose}{A logical value indicating whether to display progress
messages during the conversion process. The default value is TRUE}
}
\description{
Uses `docxtractr::convert_to_pdf()` for conversion.
}
