% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_notes.R
\name{xml_notes}
\alias{xml_notes}
\title{Get Notes from XML}
\usage{
xml_notes(file, collapse_text = TRUE, xpath = "//a:r//a:t")
}
\arguments{
\item{file}{XML file from a PPTX}

\item{collapse_text}{should text be collapsed by spaces?}

\item{xpath}{\code{xpath} to pass to [xml2::xml_find_all()]}
}
\value{
A character vector
}
\description{
Get Notes from XML
}
