% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_narrate.R
\name{ari_narrate}
\alias{ari_narrate}
\title{Generate video from slides and a script}
\usage{
ari_narrate(
  script,
  slides,
  output,
  tts_engine = text2speech::tts,
  tts_engine_args = coqui_args(),
  tts_engine_auth = text2speech::tts_auth,
  capture_method = c("vectorized", "iterative"),
  subtitles = FALSE,
  verbose = FALSE,
  cleanup = TRUE,
  ...
)
}
\arguments{
\item{script}{Either a markdown file where every paragraph will be read over
a corresponding slide, or an \code{.Rmd} file where each HTML comment will
be used for narration.}

\item{slides}{A path or URL for an HTML slideshow created with
\code{\link[rmarkdown]{rmarkdown}}, \code{xaringan}, or a
similar package.}

\item{output}{The path to the video file which will be created.}

\item{tts_engine}{The desired engine for converting text-to-speech}

\item{tts_engine_args}{List of parameters provided to the designated text-to-speech engine}

\item{tts_engine_auth}{Authentication required for the designated text-to-speech engine}

\item{capture_method}{Either \code{"vectorized"} or \code{"iterative"}.
The vectorized mode is faster though it can cause screens to repeat. If
making a video from an \code{\link[rmarkdown]{ioslides_presentation}}
you should use \code{"iterative"}.}

\item{subtitles}{Should a \code{.srt} file be created with subtitles? The
default value is \code{FALSE}. If \code{TRUE} then a file with the same name
as the \code{output} argument will be created, but with the file extension
\code{.srt}.}

\item{verbose}{print diagnostic messages.  If > 1, then more are printed}

\item{cleanup}{If \code{TRUE}, interim files are deleted}

\item{...}{Arguments that will be passed to \code{\link[webshot]{webshot}}.}
}
\value{
The output from \code{\link{ari_spin}}
}
\description{
\code{ari_narrate} creates a video from a script written in markdown and HTML
slides created with \code{\link[rmarkdown]{rmarkdown}} or a similar package.
}
\examples{
\dontrun{
ari_narrate(system.file("test", "ari_intro_script.md", package = "ari"),
            system.file("test", "ari_intro.html", package = "ari"),
            output = "test.mp4")
}
}
