% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_narrate.R
\name{ari_narrate}
\alias{ari_narrate}
\title{Create a video from slides and a script}
\usage{
ari_narrate(script, slides, output = "output.mp4", voice, ws_args = list())
}
\arguments{
\item{script}{A markdown file where every paragraph will be read over a
corresponding slide.}

\item{slides}{A path or URL for an HTML slideshow created with 
\code{\link[rmarkdown]{rmarkdown}}, \code{xaringan}, or a 
similar package.}

\item{output}{The path to the video file which will be created.}

\item{voice}{The Amazon Polly voice you want to use. See
\code{\link[aws.polly]{list_voices}}.}

\item{ws_args}{Arguments to be passed to \code{\link[webshot]{webshot}}.}
}
\description{
\code{ari_narrate} creates a video from a script written in markdown and HTML
slides created with \code{\link[rmarkdown]{rmarkdown}} or a similar package.
This function uses \href{https://aws.amazon.com/polly/}{Amazon Polly} 
via \code{\link{ari_spin}}.
}

