% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_spin.R
\name{ari_spin}
\alias{ari_spin}
\title{Ari Spin}
\usage{
ari_spin(images, paragraphs, output = "output.mp4", voice)
}
\arguments{
\item{images}{A vector of paths to images.}

\item{paragraphs}{A vector strings that will be spoken by Amazon Polly.}

\item{output}{A path to the video file which will be created.}

\item{voice}{The Amazon Polly voice you want to use. See \code{list_voices}.}
}
\description{
Given images and text, make a video
}

