% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_stitch.R
\name{ari_stitch}
\alias{ari_stitch}
\title{Ari Stitch}
\usage{
ari_stitch(images, audio, output = "output.mp4")
}
\arguments{
\item{images}{A vector of paths to images.}

\item{audio}{A list of \code{Wave}s from tuneR.}

\item{output}{A path to the video file which will be created.}
}
\description{
Given images and wavs, make a video
}

