% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_subtitles.R
\name{ari_subtitles}
\alias{ari_subtitles}
\title{Generate subtitle files for audio for video content}
\usage{
ari_subtitles(paragraphs, wavs, path, width = 42)
}
\arguments{
\item{paragraphs}{String of text}

\item{wavs}{Wave objects from tuneR}

\item{path}{Path to .srt file output}

\item{width}{Width of each subtitle}
}
\description{
Generate subtitle files for audio for video content
}
