% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_talk.R
\name{ari_talk}
\alias{ari_talk}
\title{Create spoken audio files}
\usage{
ari_talk(paragraphs, output = "output.wav", voice)
}
\arguments{
\item{paragraphs}{A vector strings that will be spoken by Amazon Polly.}

\item{output}{A path to the audio file which will be created.}

\item{voice}{The Amazon Polly voice you want to use. See 
\code{\link[aws.polly]{list_voices}} for more information about what voices
are available.}
}
\description{
A simple function for demoing how spoken text from
\href{https://aws.amazon.com/polly/}{Amazon Polly} will sound.
}
