% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg_codecs.R
\name{ffmpeg_codecs}
\alias{ffmpeg_codecs}
\alias{ffmpeg_video_codecs}
\alias{ffmpeg_audio_codecs}
\alias{ffmpeg_muxers}
\alias{ffmpeg_version}
\alias{ffmpeg_version_sufficient}
\alias{check_ffmpeg_version}
\title{Get Codecs for ffmpeg}
\usage{
ffmpeg_codecs()

ffmpeg_video_codecs()

ffmpeg_audio_codecs()

ffmpeg_muxers()

ffmpeg_version()

ffmpeg_version_sufficient()

check_ffmpeg_version()
}
\value{
A `data.frame` of codec names and capabilities
}
\description{
Get Codecs for ffmpeg
}
\examples{
\dontrun{
if (ffmpeg_version_sufficient()) {
ffmpeg_codecs()
ffmpeg_video_codecs()
ffmpeg_audio_codecs()
}
}
}
